+++ components/ad/blockerExtensionDetector/CHANGELOG.md
@@ -2,23 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 
-<a name="1.3.0"></a>
+# 1.4.0 (2020-02-10)
+
+
+### Features
+
+* Move to [@s-ui](https://github.com/s-ui) organization ([226e124](https://github.com/SUI-Components/schibsted-spain-components/commit/226e1246c2da443332e4cea1e8eacb6c5c5a092e))
+
+
+
 # 1.3.0 (2019-04-02)
 
 
 ### Features
 
-* Remove .npmignore ([68aabc3](https://github.com/SUI-Components/schibsted-spain-components/commit/68aabc3))
+* Remove .npmignore ([68aabc3](https://github.com/SUI-Components/schibsted-spain-components/commit/68aabc38867f061eb1561f5715f000a4f52dbc84))
 
 
 
-<a name="1.2.0"></a>
 # 1.2.0 (2018-10-24)
 
 
 ### Bug Fixes
 
-* deleted remove method ([78041bb](https://github.com/SUI-Components/schibsted-spain-components/commit/78041bb))
+* deleted remove method ([78041bb](https://github.com/SUI-Components/schibsted-spain-components/commit/78041bb5593c3b989fcae1c9451191d47331a163))
+
+
+
+# 1.1.0 (2018-07-27)