+++ components/button/basic/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.6.0 (2020-02-10)
+
+
+### Features
+
+* Move to [@s-ui](https://github.com/s-ui) organization ([f651051](https://github.com/SUI-Components/schibsted-spain-components/commit/f651051c96c5f6566a704bdefdadb62c11586981))
+
+
+
 # 1.5.0 (2020-01-28)